Initial commit of whiteboards. Added whiteboard, whiteboard marker (new files, so also updated polaris.dmi), created whiteboard and wb. marker icons, updated paperwork description for examine.
Whiteboards are intended to be used in areas such as research (...) for fancy information transfer. Character can read a whiteboard if it's in sight, they don't need to be right next to it. Whiteboards use fields. They also use HR to separate the text into parts, which can be erased separately. Oh and, they can be erased. The eraser is attached to the whiteboard (NOTE: I might actually make it a separate item, but for now I thought it would be too much).